refactoring stub,add networkError and add test
