+++ solutions/devsprint-1/DeliveryAppChallengeTests/Mocks/NetworkManagerStub.swift
@@ -18,17 +18,22 @@ final class NetworkManagerProtocolStub: NetworkManagerProtocol {
     
     // MARK: - Public Properties
     
-    var requestToBeReturned: Result<[Any],Error>?
+    var requestToBeReturned: Any?
     
     // MARK: - Public Functions
     
     func request<T>(_ request: NetworkRequest, completion: @escaping NetworkResult<T>) where T : Decodable {
         requestCalled = true
         requestPassed = request
-        guard let requestToBeReturned = requestToBeReturned, let requestMock = requestToBeReturned as? Result<[T],Error> else {
-            return
-        }
+//        guard let requestToBeReturned = requestToBeReturned, let requestMock = requestToBeReturned as? Result<T,Error> else {
+//            return
+//        }
+//
+//        completion(requestMock)
         
-        completion(requestMock)
+        if let requestToBeReturned = requestToBeReturned,
+            let foo = requestToBeReturned as? Result<T, Error> {
+            completion(foo)
+        }
     }
 }
